requirements.txt
django~=3.0.8
psycopg2
dj_database_url
djangorestframework~=3.11.0
PyJWT~=1.7.1
dj-config-url~=0.1.1